requirements.txt
# Core pipeline
opencv-python-headless>=4.8,<5
paddleocr>=2.7
pytesseract>=0.3
reportlab>=4.0
PyMuPDF>=1.23
Pillow>=10.0
pydantic>=2.0,<3
click>=8.0,<9
PyYAML>=6.0

# AI assist (optional — pipeline works without these)
openai>=1.0
google-genai>=1.0
anthropic>=0.30
python-dotenv>=1.0
chromadb>=0.5

# Corpus ingestion (optional — for UFLI curriculum indexing)
python-pptx>=0.6.21
playwright>=1.40

# Dev / CI
pytest>=7.0
mypy>=1.5
ruff>=0.4
types-PyYAML>=6.0
types-Pillow>=10.0
